+++ minerva/minerva/network/interface/WeightRegularizationCostFunction.h
@@ -22,7 +22,7 @@ public:
 
 public:
 	/*! \brief Run the cost function on the specified weights. */
-	virtual float getCost(const Matrix& weights) const;
+	virtual double getCost(const Matrix& weights) const;
 
 	/*! \brief Compute the gradient with respect to the weights. */
 	virtual Matrix getGradient(const Matrix& weights) const;
