+++ minerva/minerva/optimizer/interface/GPULBFGSSolver.h
@@ -21,7 +21,7 @@ public:
 	virtual ~GPULBFGSSolver();
 
 public:
-	virtual float solve(MatrixVector& inputs, 
+	virtual double solve(MatrixVector& inputs, 
 		const CostAndGradientFunction& callback);
 
 public:
